fix(user): truncate syscall numbers to 32 bits on x86-64 and aarch64
Linux kernel entry points on x86-64 (do_syscall_64) and arm64
(el0_svc_common) read the system call number as a 32-bit integer,
ignoring any values in the upper 32 bits of the register. In contrast,
RISC-V 64 (do_trap_ecall_u) bounds the full 64-bit register value.

Truncate the syscall number to 32 bits on x86-64 and aarch64 prior to
table lookup, and add tests verifying entry register width behavior.

+++ src/user/linux/abi/mod.rs
@@ -103,11 +103,13 @@ impl LinuxAbi {
         (ElfClass::Elf64, ElfData::Lsb)
     }
 
-    /// The system call with ABI number `nr`.
+    /// The system call the number register `nr` selects: x86-64 and
+    /// arm64 read the register as an `int` (`do_syscall_64`,
+    /// `el0_svc_common`), RV64 as a `long` (`do_trap_ecall_u`).
     pub fn sysno(self, nr: u64) -> Option<Sysno> {
         match self {
-            LinuxAbi::X86_64 => syscalls::x86_64_sysno(nr),
-            LinuxAbi::Aarch64 => syscalls::aarch64_sysno(nr),
+            LinuxAbi::X86_64 => syscalls::x86_64_sysno(u64::from(nr as u32)),
+            LinuxAbi::Aarch64 => syscalls::aarch64_sysno(u64::from(nr as u32)),
             LinuxAbi::Riscv64 => syscalls::riscv64_sysno(nr),
         }
     }
